+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/backups/adapters/TableBackupAdapterFactory.java
@@ -2,6 +2,7 @@ package de.mephisto.vpin.server.backups.adapters;
 
 import de.mephisto.vpin.restclient.backups.BackupType;
 import de.mephisto.vpin.restclient.frontend.TableDetails;
+import de.mephisto.vpin.server.backups.BackupSource;
 import de.mephisto.vpin.server.backups.adapters.vpa.TableBackupAdapterVpa;
 import de.mephisto.vpin.server.backups.adapters.vpa.VpaService;
 import de.mephisto.vpin.server.games.Game;
@@ -23,13 +24,13 @@ public class TableBackupAdapterFactory {
   @Autowired
   private VpaService vpaService;
 
-  public TableBackupAdapter createAdapter(@NonNull Game game) {
+  public TableBackupAdapter createAdapter(@NonNull Game game, @NonNull BackupSource backupSource) {
     BackupType backupType = systemService.getBackupType();
     TableDetails tableDetails = frontendService.getTableDetails(game.getId());
 
     switch (backupType) {
       case VPA: {
-        return new TableBackupAdapterVpa(vpaService, game, tableDetails);
+        return new TableBackupAdapterVpa(vpaService, backupSource, game, tableDetails);
       }
       default: {
         throw new UnsupportedOperationException("Unkown archive type " + backupType);